\name{iNOTE-package}
\alias{iNOTE-package}
\alias{iNOTE}
\docType{package}
\title{
\packageTitle{iNOTE}
}
\description{
\packageDescription{iNOTE}
}

\author{
\packageAuthor{iNOTE}

Maintainer: \packageMaintainer{iNOTE}
}
\references{
Chu S.H. and Huang Y-T. (2017)
Integrative genomic analysis of biological gene sets with applications in lung cancer. (Revise and Resubmit)

Huang Y-T, Lin X. (2013)
Gene set analysis using variance component tests. 
\emph{BMC Bioinformatics}. \bold{14}(1):210. PMCID: PMC3776447

Huang Y-T, Vanderweele TJ, Lin X. (2014)
Joint analysis of SNP and gene expression data in genetic association studies of complex diseases. 
\emph{The Annals of Applied Statistics}. \bold{8}(1):352–76.
}


\keyword{ iNOTE }