\name{Y}
\alias{Y}
\docType{data}
\title{
Lung Cancer Patient 1-Year Survival Status 
}
\description{Lung cancer survival status at one year after initial diagnosis (0=non-survivor, 1=survivor).
}
\usage{data("Y")}
\format{
 An R vector with length of 249.
}

\source{
  \describe{
  \item{TCGA}{http://cancergenome.nih.gov/}
  }
}

\examples{
data(Y)
str(Y)
}
\keyword{datasets}
